+++ include/Gaffer/Private/IECorePreview/LRUCache.h
@@ -115,6 +115,10 @@ class LRUCache : private boost::noncopyable
 		/// Throws if the item can not be computed.
 		Value get( const GetterKey &key );
 
+		/// Retrieves an item from the cache if it has been computed or set
+		/// previously. Throws if a previous call to `get()` failed.
+		boost::optional<Value> getIfCached( const Key &key );
+
 		/// Adds an item to the cache directly, bypassing the GetterFunction.
 		/// Returns true for success and false on failure - failure can occur
 		/// if the cost exceeds the maximum cost for the cache. Note that even
@@ -176,7 +180,7 @@ class LRUCache : private boost::noncopyable
 			//
 			// - Uncached : A boost::blank instance
 			// - Cached : The Value itself
-			// - Failed ; The exception thrown by the GetterFn
+			// - Failed : The exception thrown by the GetterFn
 			typedef boost::variant<boost::blank, Value, std::exception_ptr> State;
 
 			State state;