+++ python/GafferUI/SpreadsheetUI/_ClipboardAlgo.py
@@ -47,6 +47,8 @@ import IECore
 #
 # `plugMatrix` should be a row-major list of value plugs,
 # as returned by createPlugMatrixFromCells, ie: [ [ r1c1, ... ], [ r2c1, ... ] ]
+#
+# \note Triggers compute of the source plugs to determine value compatibility.
 def canCopyPlugs( plugMatrix ) :
 
 	if not plugMatrix :
@@ -59,21 +61,24 @@ def canCopyPlugs( plugMatrix ) :
 	if len( plugMatrix ) > 1 :
 
 		def rowData( row ) :
-			return [ __getValueAsData( cell ) for cell in row ]
+			return [ ValueAdaptor.get( cell ) for cell in row ]
 
 		columnTemplate = rowData( plugMatrix[0] )
 		for row in plugMatrix[ 1 : ] :
-			if not __dataSchemaMatches( rowData( row ), columnTemplate )  :
+			if not ValueAdaptor.dataSchemaMatches( rowData( row ), columnTemplate )  :
 				return False
 
 	return True
 
 ## Builds a 'paste-able' data for the supplied plug matrix
+# For Spreadsheet rows, the matrix should consist of a single column
+# containing the Spreadsheet.RowPlug for each row to be copied.
+# \see copyRows
 def valueMatrix( plugMatrix ) :
 
 	assert( canCopyPlugs( plugMatrix ) )
 	return IECore.ObjectVector(
-		[ IECore.ObjectVector( [ __getValueAsData( column ) for column in row ] ) for row in plugMatrix ]
+		[ IECore.ObjectVector( [ ValueAdaptor.get( column ) for column in row ] ) for row in plugMatrix ]
 	)
 
 # Returns True if the supplied object appears to be pasteable cell data
@@ -94,7 +99,7 @@ def isValueMatrix( data ) :
 		return False
 
 	for i in range( 1, len(data) ) :
-		if not __dataSchemaMatches( data[i], templateRow ) :
+		if not ValueAdaptor.dataSchemaMatches( data[i], templateRow ) :
 			return False
 
 	return True
@@ -102,7 +107,10 @@ def isValueMatrix( data ) :
 # Returns True if the supplied data can be pasted on to the supplied
 # spreadsheet cell plugs, in that the cell value types are compatible with the
 # corresponding valueMatrix.
+# \note Triggers compute of the target plugs to determine value compatibility.
 def canPasteCells( valueMatrix, plugMatrix ) :
+
+	valueMatrix = __coerceToValueMatrixIfRequired( valueMatrix )
 
 	if not isValueMatrix( valueMatrix ) :
 		return False
@@ -121,21 +129,31 @@ def canPasteCells( valueMatrix, plugMatrix ) :
 	for targetRowIndex, row in enumerate( plugMatrix ) :
 		for targetColumnIndex, cell in enumerate( row ) :
 			data = __dataForPlug( targetRowIndex, targetColumnIndex, valueMatrix )
-			if not __dataSchemaMatches( data, __getValueAsData( cell ) ) :
+			if not ValueAdaptor.canSet( data, cell ) :
 				return False
 
 	return True
 
 def pasteCells( valueMatrix, plugs, atTime ) :
 
+	valueMatrix = __coerceToValueMatrixIfRequired( valueMatrix )
+
 	assert( canPasteCells( valueMatrix, plugs ) )
 
-	for targetRowIndex, row in enumerate( plugs ) :
-		for targetColumnIndex, cell in enumerate( row ) :
-			__setValueFromData( cell, __dataForPlug( targetRowIndex, targetColumnIndex, valueMatrix ), atTime )
+	for rowIndex, row in enumerate( plugs ) :
+		for columnIndex, cell in enumerate( row ) :
+			ValueAdaptor.set( __dataForPlug( rowIndex, columnIndex, valueMatrix ), cell, atTime )
 
-# Returns True if the supplied data can be pasted as new rows, this
-# requires the target plugs columns to have matching data types.
+## Returns a value matrix for the supplied row plugs.
+def copyRows( rowPlugs ) :
+
+	return valueMatrix( [ [ row ] for row in rowPlugs ] )
+
+## Returns True if the supplied data can be pasted as new rows.
+# Columns are matched by name (and type), allowing rows to be copied
+# between Spreadsheets with different configurations. Cells in the
+# target Spreadsheet with no data will be set to the default value for
+# that column.
 def canPasteRows( data, rowsPlug ) :
 
 	if not isValueMatrix( data ) :
@@ -145,8 +163,7 @@ def canPasteRows( data, rowsPlug ) :
 	if Gaffer.MetadataAlgo.readOnly( rowsPlug ) :
 		return False
 
-	defaultsData = valueMatrix( [ rowsPlug.defaultRow().children() ] )[0]
-	return __dataSchemaMatches( data[0], defaultsData )
+	return canPasteCells( data, [ [ rowsPlug.defaultRow() ] ] )
 
 # Pastes the supplied data as new rows at the end of the supplied rows plug.
 def pasteRows( valueMatrix, rowsPlug ) :
@@ -156,7 +173,7 @@ def pasteRows( valueMatrix, rowsPlug ) :
 	# addRows currently returns None, so this is easier
 	newRows = [ rowsPlug.addRow() for _ in valueMatrix ]
 	# We know these aren't animated as we've just made them so time is irrelevant
-	pasteCells( valueMatrix, [ row.children() for row in newRows ], 0 )
+	pasteCells( valueMatrix, [ [ row ] for row in newRows ], 0 )
 
 ## Takes an arbitrary list of spreadsheet CellPlugs (perhaps as obtained from a
 # selection, which may be in a jumbled order) and groups them, ordered by row
@@ -189,72 +206,273 @@ def createPlugMatrixFromCells( cellPlugs ) :
 
 	return matrix
 
+def __coerceToValueMatrixIfRequired( data ) :
+
+	if isinstance( data, IECore.Data ) :
+		data = IECore.ObjectVector( [ IECore.ObjectVector( [ data ] ) ] )
+
+	return data
+
 # Wraps the lookup indices into the available data space
 def __dataForPlug( targetRowIndex, targetColumnIndex, data ) :
 
 	return data[ targetRowIndex % len(data) ][ targetColumnIndex % len(data[0]) ]
 
-# Returns an IECore.Data that represents either the plug value, or a hierarchy
-# of CompoundData representing the values of the plug's leaves.
-def __getValueAsData( plug ) :
+## Value Adaptors bridge plugs and data within a value matrix.
+# They allow custom extraction of plug data for copy or mis-matched types to be
+# adapted to a target plug for paste. Adaptors are registered via plug class.
+# Callers should always use the base class ValueAdaptor get/canSet/set methods.
+# Derived classes should re-implement _canSet, _set or _get as required.
+#
+# Note: At present, unless overridden, ValueAdaptors are not recursive. ie: An
+# adaptor will only be used if there is one registered for the specific plug
+# passed to get/canSet/set. Some adaptors may choose to call get/canSet/set
+# themselves to allow other registered adaptors to run for their child plugs.
+# This prohibits per-leaf/nested value adaption, but hopefully simplifies
+# understanding of when a specific adaptor may run.
+class ValueAdaptor :
 
-	if hasattr( plug, 'getValue' ) :
-		return IECore.CompoundData( { "v" : plug.getValue() } )["v"]
+	__registry = {}
 
-	return IECore.CompoundData( { child.getName() : __getValueAsData( child ) for child in plug } )
+	# \return An IECore.Data that represents either the plug value, or a hierarchy
+	# of CompoundData representing the values of the plug's leaves.
+	@staticmethod
+	def get( plug ) :
 
-def __setValueFromData( plug, data, atTime ) :
+		return ValueAdaptor.__adaptor( plug )._get( plug )
 
-	if Gaffer.MetadataAlgo.readOnly( plug ) :
-		return
+	## \return True if the supplied data can be set on the specified plug either
+	# directly, or via some transformation.
+	# \note Triggers compute of the target plugs.
+	@staticmethod
+	def canSet( data, plug ) :
 
-	if isinstance( plug, Gaffer.NameValuePlug ) :
+		return ValueAdaptor.__adaptor( plug )._canSet( data, plug )
+
+	## Sets the specified plug's value at the given time, keyframing animated values.
+	# \note This should be called from within an UndoScope.
+	@staticmethod
+	def set( data, plug, atTime ) :
+
+		ValueAdaptor.__adaptor( plug )._set( data, plug, atTime )
+
+	## \return True if the schema (ie. class hierarchy) of the two supplied
+	# objects match, regardless of their values.
+	@staticmethod
+	def dataSchemaMatches( data, otherData ) :
+
+		if type( data ) != type( otherData ) :
+			return False
+
+		if isinstance( data, ( dict, IECore.CompoundData ) ) :
+
+			if data.keys() != otherData.keys() :
+				return False
+			for a, b in zip( data.values(), otherData.values() ) :
+				if not ValueAdaptor.dataSchemaMatches( a, b ) :
+					return False
+
+		elif isinstance( data, ( list, tuple, IECore.ObjectVector ) ) :
+
+			if len( data ) != len( otherData ) :
+				return False
+			for a, b in zip( data, otherData ) :
+				if not ValueAdaptor.dataSchemaMatches( a, b ) :
+					return False
+
+		return True
+
+	@staticmethod
+	def registerAdaptor( plugType, cls ) :
+
+		ValueAdaptor.__registry[ plugType ] = cls
+
+	@classmethod
+	def _get( cls, plug ) :
+
+		if hasattr( plug, 'getValue' ) :
+			return IECore.CompoundData( { "v" : plug.getValue() } )["v"]
+
+		return IECore.CompoundData( { child.getName() : cls._get( child ) for child in plug } )
+
+	## Derived classes should take care to ensure that _canSet only returns True
+	# when _set is capable of transforming the supplied data so that it can be
+	# set on the specified plug.
+	@classmethod
+	def _canSet( cls, data, plug ) :
+
+		# Ensure we consider child plugs, eg: components of a V3fPlug
+		if Gaffer.MetadataAlgo.readOnly( plug ) :
+			return False
+		for p in Gaffer.Plug.RecursiveRange( plug ) :
+			if Gaffer.MetadataAlgo.getReadOnly( p ) :
+				return False
+
+		plugData = ValueAdaptor.get( plug )
+
+		if cls.dataSchemaMatches( data, plugData ) :
+			return True
+
+		# Support basic value embedding for NameValuePlug -> ValuePlug
+		if isinstance( data, IECore.CompoundData ) and "value" in data :
+			return ValueAdaptor.dataSchemaMatches( data[ "value" ], plugData )
+
+		return False
+
+	## Derived classes should re-implement _canSet to match any custom logic added here.
+	@classmethod
+	def _set( cls, data, plug, atTime ) :
+
+		if isinstance( data, IECore.CompoundData ) :
+			# Only perform schema check for compound data types, to allow value coercion for basic types
+			if not ValueAdaptor.dataSchemaMatches( data, ValueAdaptor.get( plug ) ) and "value" in data :
+				data = data[ "value" ]
+
+		if hasattr( plug, 'setValue' ) :
+			ValueAdaptor._setOrKeyValue( plug, data, atTime )
+		else :
+			for childName, childData in data.items() :
+				cls._set( childData, plug[ childName ], atTime )
+
+	@staticmethod
+	def _setOrKeyValue( plug, value, atTime ) :
+
+		if hasattr( value, 'value' ) :
+			value = value.value
+
+		if Gaffer.Animation.isAnimated( plug ) :
+			curve = Gaffer.Animation.acquire( plug )
+			if not Gaffer.MetadataAlgo.readOnly( curve ) :
+				curve.addKey( Gaffer.Animation.Key( atTime, value, Gaffer.Animation.Type.Linear ) )
+		elif plug.settable() :
+			plug.setValue( value )
+
+	@staticmethod
+	def __adaptor( plug ) :
+
+		return ValueAdaptor.__registry.get( type( plug ), ValueAdaptor )
+
+class NameValuePlugValueAdaptor( ValueAdaptor ) :
+
+	@classmethod
+	def _canSet( cls, data, nameValuePlug ) :
+
+		if isinstance( data, IECore.CompoundData ) and "value" in data :
+			valueData = data[ "value" ]
+			if "enabled" in nameValuePlug :
+				if "enabled" in data and not ValueAdaptor.canSet( data[ "enabled" ], nameValuePlug[ "enabled" ] ) :
+					return False
+		else :
+			# Allow simple data to be pasted onto the value plug
+			valueData = data
+
+		return ValueAdaptor.canSet( valueData, nameValuePlug[ "value" ] )
+
+	@classmethod
+	def _set( cls, data, nameValuePlug, atTime ) :
 
 		# We should _never_ set the name of an NVP as it's not exposed in the
 		# UI. It's only there as it simplifies things greatly if we don't
 		# special case plugs in the general case in the Spreadsheet. If we did,
 		# and you pasted across columns, it would duplicate the plug name,
 		# which can have catastrophic results for nodes such as StandardOptions.
-		for child in plug.children() :
-			if child.getName() != "name" :
-				__setValueFromData( child, data[ child.getName() ], atTime )
 
-	elif hasattr( plug, 'setValue' ) :
+		if isinstance( data, IECore.CompoundData ) and "value" in data :
+			valueData = data[ "value" ]
+			enabledData = data[ "enabled" ] if "enabled" in data else None
+		else :
+			# Allow simple data to be pasted onto the value plug
+			valueData = data
+			enabledData = None
 
-		if hasattr( data, 'value' ) :
-			data = data.value
+		ValueAdaptor.set( valueData, nameValuePlug[ "value" ], atTime )
 
-		if Gaffer.Animation.isAnimated( plug ) :
-			curve = Gaffer.Animation.acquire( plug )
-			if not Gaffer.MetadataAlgo.readOnly( curve ) :
-				curve.addKey( Gaffer.Animation.Key( atTime, data, Gaffer.Animation.Type.Linear ) )
-		elif plug.settable() :
-			plug.setValue( data )
+		if "enabled" in nameValuePlug and enabledData is not None :
+			ValueAdaptor.set( enabledData, nameValuePlug[ "enabled" ], atTime )
 
-	else :
+ValueAdaptor.registerAdaptor( Gaffer.NameValuePlug, NameValuePlugValueAdaptor )
 
-		for childName, childData in data.items() :
-			__setValueFromData( plug[ childName ], childData, atTime )
+class CellPlugValueAdaptor( ValueAdaptor ) :
 
-def __dataSchemaMatches( data, otherData ) :
+	@classmethod
+	def _canSet( cls, data, cellPlug ) :
 
-	if type( data ) != type( otherData ) :
-		return False
+		enabledData, valueData = CellPlugValueAdaptor.__enabledAndValueData( data )
 
-	if isinstance( data, ( dict, IECore.CompoundData ) ) :
-
-		if data.keys() != otherData.keys() :
+		if enabledData is not None and not ValueAdaptor.canSet( enabledData, cellPlug.enabledPlug() ) :
 			return False
-		for a, b in zip( data.values(), otherData.values() ) :
-			if not __dataSchemaMatches( a, b ) :
-				return False
 
-	elif isinstance( data, ( list, tuple, IECore.ObjectVector ) ) :
+		return ValueAdaptor.canSet( valueData, cellPlug[ "value" ] )
 
-		if len( data ) != len( otherData ) :
+	@classmethod
+	def _set( cls, data, cellPlug, atTime ) :
+
+		enabledData, valueData = CellPlugValueAdaptor.__enabledAndValueData( data )
+
+		ValueAdaptor.set( valueData, cellPlug[ "value" ], atTime )
+
+		# Set enabled state last, such that when copying from a cell that doesn't adopt
+		# an enabled plug, to one that does, the final 'enabled' state matches.
+		if enabledData is not None :
+			ValueAdaptor.set( enabledData, cellPlug.enabledPlug(), atTime )
+
+	@staticmethod
+	def __enabledAndValueData( data ) :
+
+		enabledData = None
+
+		if isinstance( data, IECore.CompoundData ) and "value" in data :
+			valueData = data[ "value" ]
+			if "enabled" in data :
+				enabledData = data[ "enabled" ]
+			elif "enabled" in valueData :
+				enabledData = valueData[ "enabled" ]
+		else :
+			valueData = data
+
+		return enabledData, valueData
+
+ValueAdaptor.registerAdaptor( Gaffer.Spreadsheet.CellPlug, CellPlugValueAdaptor )
+
+class FloatPlugValueAdaptor( ValueAdaptor ) :
+
+	@classmethod
+	def _canSet( cls, data, plug ) :
+
+		# FloatPlug.setValue will take care of this for us
+		return isinstance( data, ( IECore.FloatData, IECore.IntData ) )
+
+ValueAdaptor.registerAdaptor( Gaffer.FloatPlug, FloatPlugValueAdaptor )
+
+## Allows RowPlugs to be copy/pasted across different column configurations,
+# matching by the column names.
+class RowPlugValueAdaptor( ValueAdaptor ) :
+
+	@classmethod
+	def _canSet( cls, data, rowPlug ) :
+
+		if not isinstance( data, IECore.CompoundData ) or set( data.keys() ) != { "name", "enabled", "cells" } :
 			return False
-		for a, b in zip( data, otherData ) :
-			if not __dataSchemaMatches( a, b ) :
-				return False
 
-	return True
+		cellData = data[ "cells" ]
+		matchingCells = [ cell for cell in rowPlug[ "cells" ].children() if cell.getName() in cellData ]
+		if not matchingCells :
+			return False
+
+		return all( [ ValueAdaptor.canSet( cellData[ cell.getName() ], cell ) for cell in matchingCells ] )
+
+	@classmethod
+	def _set( cls, rowData, rowPlug, atTime ) :
+
+		# Name/enabled
+		for plugName in ( "name", "enabled" ) :
+			ValueAdaptor.set( rowData[ plugName ], rowPlug[ plugName ], atTime )
+
+		# Cells
+		cellData = rowData[ "cells" ]
+		for cell in rowPlug[ "cells" ].children() :
+			data = cellData.get( cell.getName(), None )
+			if data is not None :
+				ValueAdaptor.set( data, cell, atTime )
+
+ValueAdaptor.registerAdaptor( Gaffer.Spreadsheet.RowPlug, RowPlugValueAdaptor )