+++ Changes.md
@@ -196,8 +196,9 @@ Improvements
 - Spreadsheet :
   - Enabled cell selection, editing is now achieved with a double-click.
   - Added menu items to cells & defaults sections to copy/paste the enabled state and value(s) of selected cells.
-  - Added menu items to the row names section to copy/paste or delete the selected rows.
+  - Added menu items to the row names section to disable, copy/paste or delete the selected rows.
   - Added menu items to edit the values for all selected cells of the same type simultaneously.
+  - Added menu items to the section chooser to quickly select a specific section.
 - Stats app : Added `-serialise` argument to measure the time taken to serialise the script.
 
 Fixes
@@ -207,11 +208,15 @@ Fixes
 - SplinePlug :
   - Fixed bugs affecting default values. All child plugs are now at their default values following construction and following a call to either `setToDefault()` or `resetDefault()`.
   - Fixed bug that prevented a spline from being promoted to a Box if it had a non-default number of points. This also affected the use of a spline as the input to an Expression.
+- LRUCache : Fixed bug which could cause hangs during scene generation (#4016).
+- ArnoldShader : Moved the toon shader's `rim_light_tint` and `aov_prefix` parameters to appropriate sections in the UI.
+- Spreadsheet : Fixed bug that caused sections to overflow the available space.
 
 API
 ---
 
 - Catalogue : Added `gaffer:isRendering` metadata, set to `True` if the viewed image is still receiving data from a display driver.
+- LRUCache : Added `getIfCached()` method.
 
 0.58.5.2 (relative to 0.58.5.1)
 ========